chore: homogenize experience when adding NFT and adding token
